pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
pattern = "default-unprefixed"

[tool.poetry.dependencies]
python = "^3.6"

# Python lack of functionalities from future versions
importlib-metadata = { version = "*", python = "<3.8" }

neptune-client = ">=0.16.16"
tensorflow = ">2.0.0"

# dev
pre-commit = { version = "*", optional = true }
pytest = { version = ">=5.0", optional = true }
pytest-cov = { version = "2.10.1", optional = true }
pydot = { version = "*", optional = true }

[tool.poetry.extras]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pydot",
]

[tool.poetry]
authors = ["neptune.ai <contact@neptune.ai>"]
description = "Neptune.ai tensorflow-keras integration library"
repository = "https://github.com/neptune-ai/neptune-tensorflow-keras"
homepage = "https://neptune.ai/"
documentation = "https://docs.neptune.ai/integrations-and-supported-tools/model-training/tensorflow-keras"
include = ["CHANGELOG.md"]
license = "Apache License 2.0"
name = "neptune-tensorflow-keras"
readme = "README.md"
version = "0.0.0"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = [
    "MLOps",
    "ML Experiment Tracking",
    "ML Model Registry",
    "ML Model Store",
    "ML Metadata Store",
]
packages = [
    { include = "neptune_tensorflow_keras", from = "src" },
]

[tool.poetry.urls]
"Tracker" = "https://github.com/neptune-ai/neptune-tensorflow-keras/issues"
"Documentation" = "https://docs.neptune.ai/integrations-and-supported-tools/model-training/tensorflow-keras"

[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
force_grid_wrap = 2

[tool.flake8]
max-line-length = 120
extend-ignore = "E203"
